@@ -1,11 +1,10 @@
 // routes/card.js
-const express = require("express");
+import express from "express";
 const router = express.Router();
-const drawController = require("../controllers/drawController");
-const drawTime = require("../middleware/drawTimeValidation");
-const authenticateToken = require("../middleware/authenticateToken"); 
+import { getDrawCards } from "../controllers/drawController.js";
+import drawTime from "../middleware/drawTimeValidation.js";
+import authenticateToken from "../middleware/authenticateToken.js";
 
+router.get("/draw", authenticateToken, drawTime, getDrawCards); // Vérifie le token, vérifie si l'utilisateur a déjà tiré une carte aujourd'hui et renvoie les cartes tirées par l'utilisateur
 
-router.get("/draw", authenticateToken, drawTime, drawController.getDrawCards); // Vérifie le token, vérifie si l'utilisateur a déjà tiré une carte aujourd'hui et renvoie les cartes tirées par l'utilisateur
-
-module.exports = router;
+export default router;
